+++ lib/features/hot_desk_booking/models/workspace.g.dart
@@ -1,0 +1,29 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'workspace.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$WorkspaceImpl _$$WorkspaceImplFromJson(Map<String, dynamic> json) =>
+    _$WorkspaceImpl(
+      id: json['id'] as String,
+      name: json['name'] as String,
+      isActive: json['isActive'] as bool,
+      createdAt: const TimestampConverter().fromJson(
+        json['createdAt'] as Timestamp,
+      ),
+      updatedAt: const TimestampConverter().fromJson(
+        json['updatedAt'] as Timestamp,
+      ),
+    );
+
+Map<String, dynamic> _$$WorkspaceImplToJson(_$WorkspaceImpl instance) =>
+    <String, dynamic>{
+      'id': instance.id,
+      'name': instance.name,
+      'isActive': instance.isActive,
+      'createdAt': const TimestampConverter().toJson(instance.createdAt),
+      'updatedAt': const TimestampConverter().toJson(instance.updatedAt),
+    };